refactor(map): moved findSegments to lib/map/segments and update import
