+++ frontend/src/lib/map/segments/findSegments.ts
